pyproject.toml
[project]
name = "django-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "ddiyoung",email = "wsw3998@sju.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
